requirements.txt
av==14.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.6.0
diskcache==5.6.3
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
numpy==2.2.5
onnxruntime==1.21.1
packaging==25.0
protobuf==6.30.2
pycparser==2.22
Pygments==2.19.1
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scipy==1.15.2
shellingham==1.5.4
sounddevice==0.5.1
sympy==1.13.3
tokenizers==0.21.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.13.2
urllib3==2.4.0
